+++ packages/storybook/stories/Accessibility.stories.mdx
@@ -1,0 +1,10 @@
+import { Meta } from '@storybook/addon-docs/blocks';
+import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/addon-docs/blocks';
+
+<Meta title="Accessibility" />
+
+# Localization
+
+We are currently improving this area on how to drive accessibility in your communications experience with the Azure Communication Services UI Toolkit
+
+## TBD
