+++ packages/storybook/stories/Composite/Handlers.stories.mdx
@@ -1,8 +1,7 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/addon-docs/blocks';
+import { STATEFUL_CLIENT_PREFIX } from './../constants';
 
-<Meta title="Composites/Custom User Data" />
-
-# Custom User Data
+<Meta title={`${STATEFUL_CLIENT_PREFIX}/Handlers`} />
 
 TBD
