+++ packages/storybook/stories/Examples/Participant List/snippets/BasicParticipantList.snippet.tsx
