+++ packages/storybook/stories/Examples/Participant List/snippets/InteractiveParticipantList.snippet.tsx
