+++ packages/storybook/stories/GettingStarted.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/addon-docs/blocks';
 
-<Meta title="Introduction" />
+<Meta title="Getting Started" />
 
 # Azure Communication Services - UI Toolkit
