+++ packages/storybook/stories/HowToUse.stories.mdx
@@ -1,0 +1,8 @@
+import { Meta } from '@storybook/addon-docs/blocks';
+import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/addon-docs/blocks';
+
+<Meta title={`How to use Storybook`} />
+
+# How to use storybook
+
+TBD
