+++ packages/storybook/stories/Localization.stories.mdx
@@ -5,4 +5,6 @@ import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/a
 
 # Localization
 
-TBD
+We are currently improving this feature and will create additional documentation on how to extend in your localized experience into your UI Toolkit powered communications implementation.
+
+## TBD
