+++ packages/storybook/stories/Personalization.stories.mdx
@@ -1,0 +1,10 @@
+import { Meta } from '@storybook/addon-docs/blocks';
+import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/addon-docs/blocks';
+
+<Meta title="Personalization" />
+
+# Personalization
+
+We are currently improving additional documentation on how to personalize your communications experience with the Azure Communication Services UI Toolkit.
+
+## TBD
